Release/student management

+++ Frontend/Cantine/src/app/student/dashbord/student-dashboard.service.ts
@@ -9,16 +9,20 @@ import {ErrorResponse} from "../../sharedmodule/models/ErrorResponse";
 import {catchError, throwError} from "rxjs";
 import {ExceptionDialogComponent} from "../../sharedmodule/dialogs/exception-dialog/exception-dialog.component";
 import Malfunctions from "../../sharedmodule/functions/malfunctions";
+import {IConstantsURL} from "../../sharedmodule/constants/IConstantsURL";
+import {DialogErrors} from "../../sharedmodule/functions/dialogueErrors";
 
 @Injectable()
 export class StudentDashboardService {
 
-    private BASIC_ENDPOINT = "http://localhost:8080/cantine/student/";
+    private BASIC_ENDPOINT = "http://localhost:8080/cantine/user/student/";
 
     private STUDENT_SIGN_UP_URL = this.BASIC_ENDPOINT + 'signUp';
     private GET_ALL_STUDENT_CLASS = this.BASIC_ENDPOINT + 'getAllStudentClass';
 
     private UPDATE_STUDENT_INFO = this.BASIC_ENDPOINT + 'update/studentInfo'
+
+    private dialog = new DialogErrors(this.matDialog);
 
     constructor(private httpClient: HttpClient, private matDialog: MatDialog, private router: Router) {
     }
@@ -26,53 +30,74 @@ export class StudentDashboardService {
     updateStudent(student: FormData) {
         let token = Malfunctions.getTokenFromLocalStorage();
         const headers = new HttpHeaders().set('Authorization', token);
-       return this.httpClient.put<NormalResponse>(this.UPDATE_STUDENT_INFO, student, {
+        return this.httpClient.post<NormalResponse>(this.UPDATE_STUDENT_INFO, student, {
             headers: headers,
-        }).pipe(catchError((error) => this.handleError(error)));
+        }).pipe(catchError((error) => this.handleUpdateStudentErrors(error)));
     }
 
     signUpStudent(student: FormData) {
         return this.httpClient.post<NormalResponse>(this.STUDENT_SIGN_UP_URL, student).pipe(
-            catchError((error) => this.handleError(error))
+            catchError((error) => this.handleSignUpErrors(error))
         );
     }
 
     getAllStudentClass() {
+        let token = Malfunctions.getTokenFromLocalStorage();
         return this.httpClient.get<StudentClass[]>(this.GET_ALL_STUDENT_CLASS);
     }
 
-    private handleError(error: HttpErrorResponse) {
+    private handleSignUpErrors(error: HttpErrorResponse) {
         const errorObject = error.error as ErrorResponse;
         let errorMessage = errorObject.exceptionMessage;
 
-        if (error.status == HttpStatusCode.BadRequest || error.status == HttpStatusCode.NotAcceptable || error.status == HttpStatusCode.Conflict || error.status == HttpStatusCode.NotFound) {
-            this.openDialog(errorMessage, error.status);
-        } else {
-            this.openDialog(" Une  Erreur  Inconnue  c'est produite ", error.status);
+        if (errorMessage == undefined) {
+            localStorage.clear();
+            this.dialog.openDialog("Une erreur s'est produite Veuillez réessayer plus tard");
+            this.router.navigate([IConstantsURL.HOME_URL]).then(window.location.reload);
+            return throwError(() => new Error(errorMessage));
         }
-        this.router.navigate(['cantine/home']).then(
-            () => {
-                localStorage.clear()
-            }
-        );
-
+        if (error.status == HttpStatusCode.BadRequest) {
+            errorMessage = "Veuillez vérifier les informations saisies";
+            this.dialog.openDialog(errorMessage);
+        } else if (error.status == HttpStatusCode.Conflict) {
+            errorMessage = "Cet utilisateur existe déjà";
+            this.dialog.openDialog(errorMessage);
+        } else if (error.status == HttpStatusCode.NotAcceptable) {
+            errorMessage = "Format d'image non valide";
+            this.dialog.openDialog(errorMessage);
+        } else {
+            errorMessage = "Une erreur s'est produite Veuillez réessayer plus tard";
+            this.dialog.openDialog(errorMessage);
+        }
         return throwError(() => new Error(errorMessage));
-
     }
 
 
-    private openDialog(message: string, httpError: HttpStatusCode): void {
-        const result = this.matDialog.open(ExceptionDialogComponent, {
-            data: {message: message},
-            width: '40%',
-        });
+    private handleUpdateStudentErrors(error: HttpErrorResponse) {
+        const errorObject = error.error as ErrorResponse;
+        let errorMessage = errorObject.exceptionMessage;
 
-        result.afterClosed().subscribe((confirmed: boolean) => {
-
-            /* TODO  remove THE  Token    and  redirect  to  the  Main  page */
-            //this.router.navigate(['/cantine/home'] , { queryParams: { reload: 'true' } });
-
-        });
-
+        if (errorMessage == undefined) {
+            localStorage.clear();
+            this.dialog.openDialog("Une erreur s'est produite Veuillez réessayer plus tard");
+            this.router.navigate([IConstantsURL.HOME_URL]).then(window.location.reload);
+            return throwError(() => new Error(errorMessage));
+        }
+        if (error.status == HttpStatusCode.BadRequest) {
+            errorMessage = "Veuillez vérifier les informations saisies";
+            this.dialog.openDialog(errorMessage);
+        } else if (error.status == HttpStatusCode.NotAcceptable) {
+            errorMessage = "Format d'image non valide";
+            this.dialog.openDialog(errorMessage);
+        } else if (error.status == HttpStatusCode.Unauthorized) {
+            this.router.navigate([IConstantsURL.SIGN_IN_URL]).then(window.location.reload);
+            localStorage.clear()
+        } else {
+            errorMessage = "Une erreur s'est produite Veuillez réessayer plus tard";
+            this.dialog.openDialog(errorMessage);
+            localStorage.clear();
+            this.router.navigate([IConstantsURL.HOME_URL]).then(window.location.reload);
+        }
+        return throwError(() => new Error(errorMessage));
     }
 }
